#DbChema: Change Schema column type of the Datatypes table
The type was changed from jsonb to json, because Postgres rearrange fields of Json file when it is inserted.
For parsing sensor output raw data we need to know exact order of the fields.
Nevertheless, there are no negative side effects for such column data type changing, because there are no need in conditional queries for schema column and therefore there are no need to index the column.